Rename /server to /writer
This is mainly because "React Server Server" is weird so we need another
dimension.
