--- packages/react-server-dom-webpack/npm/server.js
@@ -1,3 +1,0 @@
-'use strict';
-
-module.exports = require('./server.node');
