+++ packages/react-server-dom-webpack/npm/writer.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('./writer.node.server');
