+++ packages/react-server-dom-webpack/writer.browser.server.js
